Project2/requirements.txt
streamlit
openai>=1
python-dotenv
pydantic>=2
git+https://github.com/microsoft/TinyTroupe.git